requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
Django==3.0.7
django-bootstrap4==1.1.1
django-jet==1.0.7
django-rest-framework==0.1.0
django-widget-tweaks==1.4.5
google-api-python-client==1.4.1
httplib2==0.18.0
oauth2client==4.1.3
Pillow==8.1.0
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pytz==2019.3
rsa==4.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
uritemplate==3.0.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
django-cors-headers==3.6.0
https://github.com/Barukimang/django-jet/archive/dev.zip
